#1168 mv profile-node tests into own file
